requirements.txt
pycryptodome
pyqt5
pycryptodome